A fuckton of stuff
Modularized shooter, expanded motor test to function with servo, concept of an auto mode, implemented everything

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AdvancedMotorTest.java
@@ -20,7 +20,7 @@ public class AdvancedMotorTest extends LinearOpMode {
 
     int[] directions = new int[4];
     int selectedMotor = 0;
-    float step = 1;
+    double step = 1;
     boolean previousInputDown = false;
     
     Arrays.fill(directions,0);
